# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server OA integrator.",
        "title": "Swagger OA Integrator API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/transactions/integrator": {
            "get": {
                "description": "To get all transactions made through OA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get all OA logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LPN",
                        "name": "lpn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Integrator Name",
                        "name": "integratorName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "Start At",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "End At",
                        "name": "endAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/transactions/logs": {
            "get": {
                "description": "To check overall system health",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get all logs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "Before",
                        "name": "endAt",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "After",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fields",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/transactions/oa": {
            "get": {
                "description": "To get all transactions made through OA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get all OA logs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "Start At",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "End At",
                        "name": "endAt",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exit Lane",
                        "name": "exitLane",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entry Lane",
                        "name": "entryLane",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Licence Plate Number",
                        "name": "lpn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Facility",
                        "name": "facility",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/transactions/oa-latest": {
            "get": {
                "description": "To get all latest update transactions made through OA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get latest update on OA logs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "Start At",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "dateTime",
                        "description": "End At",
                        "name": "endAt",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        }
    },
...
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}